improves cli experience by adding input validation

+++ lib/helpers.py
@@ -5,26 +5,30 @@ from models.product import Product
 ### STORE OPERATIONS ###
 def add_new_store():
     name = input("Enter the name of the new store: ")
-    location = input("Enter the location of the new store: ")
-    if name and location:
-        try:
-            store = Store.create(name, location)
-            return store
-        except Exception as exc:
-            print(f"\t>>>> Error: ", exc, "<<<<", "\n")
+    if name and not name.isdigit():
+        location = input("Enter the location of the new store: ")
+        if location and not location.isdigit():
+            try:
+                store = Store.create(name, location)
+                print(f"\n{store} has been added successfully.")
+            except Exception as exc:
+                print(f"\t>>>> Error: ", exc, "<<<<", "\n")
+        else:
+            print(f"\t>>>> Input cannot be blank <<<<")
     else:
         print(f"\t>>>> Input cannot be blank <<<<")
 
 def remove_store():
     id_ = input("Enter the id of the store: ")
-    if id_.isdigit():
+    if id_.isdigit() and int(id_) > 0:
         store = Store.find_by_id(int(id_))
         if store:
             choice = input(f"Are you sure you want to remove {store}? (y/n): ")
             if choice.lower() == "y":
+                print(f"{store} has been removed")
                 store.delete()
             elif choice.lower() == "n":
-                print(f"\nOperation cancelled")
+                print(f"Operation cancelled")
             else:
                 print("\t>>>> Invalid choice <<<<")
         else:
@@ -34,15 +38,17 @@ def remove_store():
 
 def update_store():
     id_ = input("Enter the store id: ")
-    if id_.isdigit():
+    if id_.isdigit() and int(id_) > 0:
         store = Store.find_by_id(int(id_))
         if store:
             try:
+                old_store = store
                 name = input("Enter the new store name: ")
                 store.name = name
                 location = input("Enter the new store location: ")
                 store.location = location
                 store.update()
+                print(f"{old_store} has been updated to {store}")
             except Exception as exc:
                 print(f"\t>>>> Error: ", exc, "<<<<", "\n")
         else:
@@ -52,10 +58,10 @@ def update_store():
 
 def find_store_by_id():
     id_ = input("Enter the store id: ")
-    if id_.isdigit():
+    if id_.isdigit() and int(id_) > 0:
         store = Store.find_by_id(int(id_))
         if store:
-            print(f"\n{store}\n")
+            print(f"{store}")
         else:
             print(f"\t>>>> Store with id {id_} not found <<<<")
     else:
@@ -63,18 +69,18 @@ def find_store_by_id():
 
 def find_store_by_name():
     name = input("Enter the store name: ")
-    if name:
-        store = Store.find_by_name(name)
+    if name and not name.isdigit():
+        store = Store.find_by_name(name.title())
         if store:
-            print(f"\n{store}\n")
+            print(f"{store}")
         else:
             print(f"\t>>>> Store with the name {name} not found <<<<")
     else:
-        print(f"\t>>>> Input cannot be blank <<<<")
+        print(f"\t>>>> Input must be a string <<<<")
 
 def find_store_by_location():
     location = input("Enter the location: ")
-    if location:
+    if location and not location.isdigit():
         stores = Store.find_in_location(location.title())
         if stores:
             for store in stores:
@@ -82,7 +88,7 @@ def find_store_by_location():
         else:
             print(f"\t>>>> No stores found in {location} <<<<")
     else:
-        print(f"\t>>>> Input cannot be blank <<<<")
+        print(f"\t>>>> Input must be a string <<<<")
 
 def view_all_stores():
     stores = Store.get_all()
@@ -91,9 +97,9 @@ def view_all_stores():
 
 def view_all_store_departments_id():
     id_ = input("Enter the store id: ")
-    if id_.isdigit():
+    if id_.isdigit() and int(id_) > 0:
         departments = Department.find_by_store_id(int(id_))
-        if departments:
+        if departments and int(id_) > 0:
             for department in departments:
                 print(department)
         else:
@@ -103,7 +109,7 @@ def view_all_store_departments_id():
 
 def view_all_store_departments_name():
     name = input("Enter the store name: ")
-    if name:
+    if name and not name.isdigit():
         departments = Department.find_by_store_name(name.title())
         if departments:
             for department in departments:
@@ -111,29 +117,38 @@ def view_all_store_departments_name():
         else:
             print(f"No department with the name {name} was found")
     else:
-        print(f"\t>>>> Input cannot be blank <<<<")
+        print(f"\t>>>> Input must be a string <<<<")
 
 ### DEPARTMENT OPERATIONS ###
 def new_department():
     name = input("Enter the name of the new department: ")
-    description = input("Enter the description: ")
-    store_id = input("Enter the store id: ")
-    if name and description and store_id.isdigit():
-        department = Department.create(name, description, int(store_id))
-        return department
+    if name and not name.isdigit():
+        description = input("Enter the description: ")
+        if description and not description.isdigit():
+            store_id = input("Enter the store id: ")
+            if store_id.isdigit():
+                department = Department.create(name, description, int(store_id))
+                print(department)
+            else:
+                print(f"\t>>>> Input must be a valid integer <<<<")
+        else:
+            print(f"\t>>>> Input must be a string <<<<")
     else:
-        print(f"\t>>>> Input cannot be blank and store ID must be an integer <<<<")
+        print(f"\t>>>> Input must be a string <<<<")
+
+        
 
 def delete_department():
     id_ = input("Enter the department id: ")
-    if id_.isdigit():
+    if id_.isdigit() and int(id_) > 0:
         department = Department.find_by_id(int(id_))
         if department:
             choice = input(f"Are you sure you want to remove {department}? (y/n): ")
             if choice.lower() == "y":
+                print(f"{department} has been removed")
                 department.delete()
             elif choice.lower() == "n":
-                print(f"\nOperation cancelled")
+                print(f"Operation cancelled")
             else:
                 print("\t>>>> Invalid choice <<<<")
         else:
@@ -143,10 +158,10 @@ def delete_department():
 
 def find_department_by_id():
     id_ = input("Enter the department id: ")
-    if id_.isdigit():
+    if id_.isdigit() and int(id_) > 0:
         department = Department.find_by_id(int(id_))
         if department:
-            print(f"\n{department}\n")
+            print(f"{department}")
         else:
             print(f"\t>>>> Department with id {id_} not found <<<<")
     else:
@@ -154,7 +169,7 @@ def find_department_by_id():
 
 def find_department_by_name():
     name = input("Enter the department name: ")
-    if name:
+    if name and not name.isdigit():
         departments = Department.find_by_name(name.title())
         if departments:
             for department in departments:
@@ -162,7 +177,7 @@ def find_department_by_name():
         else:
             print(f"\t>>>> Department with the name {name} not found <<<<")
     else:
-        print(f"\t>>>> Input cannot be blank <<<<")
+        print(f"\t>>>> Input must be a string <<<<")
 
 def view_all_departments():
     departments = Department.get_all()
@@ -171,7 +186,7 @@ def view_all_departments():
 
 def update_department():
     id_ = input("Enter the department id: ")
-    if id_.isdigit():
+    if id_.isdigit() and int(id_) > 0:
         department = Department.find_by_id(int(id_))
         if department:
             try:
@@ -192,25 +207,36 @@ def update_department():
 ### PRODUCT OPERATIONS ###
 def new_product():
     name = input("Enter the name of the new product: ")
-    description = input("Enter the description: ")
-    quantity = input("Enter the quantity of the product: ")
-    department_id = input("Enter the department id: ")
-    if name and description and quantity.isdigit() and department_id.isdigit():
-        try:
-            product = Product.create(name, description, int(quantity), int(department_id))
-            return product
-        except Exception as exc:
-            print(f"\t>>>> Error: ", exc, "<<<<", "\n")
+    if name and not name.isdigit():
+        description = input("Enter the product type: ")
+        if description and not description.isdigit():
+            quantity = input("Enter the quantity of the product: ")
+            if quantity.isdigit():
+                department_id = input("Enter the department id: ")
+                if department_id.isdigit():
+                    try:
+                        product = Product.create(name, description, int(quantity), int(department_id))
+                        print(product)
+                    except Exception as exc:
+                        print(f"\t>>>> Error: ", exc, "<<<<", "\n")
+                else:
+                    print(f"\t>>>> Input must be a valid integer <<<<")
+            else:
+                print(f"\t>>>> Input must be a valid integer <<<<")
+        else:
+            print(f"\t>>>> Input must be a string <<<<")
     else:
-        print(f"\t>>>> Input cannot be blank and IDs must be integers <<<<")
+        print(f"\t>>>> Input must be a string <<<<")
+
 
 def delete_product():
     id_ = input("Enter the product id: ")
-    if id_.isdigit():
+    if id_.isdigit() and int(id_) > 0:
         product = Product.find_by_id(int(id_))
         if product:
             choice = input(f"Are you sure you want to remove {product}? (y/n): ")
             if choice.lower() == "y":
+                print(f"\n{product} has been removed")
                 product.delete()
             elif choice.lower() == "n":
                 print(f"\nOperation cancelled")
@@ -223,19 +249,20 @@ def delete_product():
 
 def update_product():
     id_ = input("Enter the product id: ")
-    if id_.isdigit():
+    if id_.isdigit() and int(id_) > 0:
         product = Product.find_by_id(int(id_))
         if product:
             try:
                 name = input("Enter the new product name: ")
                 product.name = name
                 type_ = input("Enter the new product type: ")
-                product.type = type_
+                product.product_type = type_
                 quantity = input("Enter the product quantity: ")
                 product.quantity = int(quantity) if quantity.isdigit() else product.quantity
                 department_id = input("Enter the new department id: ")
                 product.department_id = int(department_id) if department_id.isdigit() else product.department_id
                 product.update()
+                print(f"{product} has been updated")
             except Exception as exc:
                 print(f"\t>>>> Error: ", exc, "<<<<", "\n")
         else:
@@ -245,10 +272,10 @@ def update_product():
 
 def find_product_by_id():
     id_ = input("Enter the product id: ")
-    if id_.isdigit():
+    if id_.isdigit() and int(id_) > 0:
         product = Product.find_by_id(int(id_))
         if product:
-            print(f"\n{product}\n")
+            print(f"{product}")
         else:
             print(f"\t>>>> Product with id {id_} not found <<<<")
     else:
@@ -273,7 +300,7 @@ def view_all_products():
 
 def view_all_products_in_department_id():
     id_ = input("Enter the department id: ")
-    if id_.isdigit():
+    if id_.isdigit() and int(id_) > 0:
         products = Product.get_all_in_department_id(int(id_))
         if products:
             for product in products:
@@ -285,29 +312,36 @@ def view_all_products_in_department_id():
 
 def view_all_products_in_department_name():
     name = input("Enter the department name: ")
-    if name:
+    if name and not name.isdigit():
         products = Product.get_all_in_department_name(name.title())
         if products:
+            products_by_store = {}
             for product in products:
                 department = Department.find_by_id(product.id)
                 store = Store.find_by_id(department.store_id)
-                print(f"\n{store.name}")
-                print(f"{product}")
+                if store.name not in products_by_store:
+                    products_by_store[store.name] = []
+                products_by_store[store.name].append(product)
+            
+            for store_name, products in products_by_store.items():
+                print(f"\n{store_name}")
+                for product in products:
+                    print(product)
         else:
             print(f"\t>>>> Department with the name {name} not found <<<<")
     else:
-        print(f"\t>>>> Input cannot be blank <<<<")
+        print(f"\t>>>> Input must be a string <<<<")
+
 
 def products_in_store_id():
     id_ = input("Enter Store ID: ")
-    if id_.isdigit():
+    if id_.isdigit() and int(id_) > 0:
         departments = Department.find_by_store_id(int(id_))
         if departments:
             store = Store.find_by_id(int(id_))
             print(f"\n{store.name}")
             for department in departments:
-                print(f"{department.name}")
-                print(f"{department.id}")
+                print(f"\n{department.name}")
                 products = Product.get_all_in_department_id(department.id)
                 for product in products:
                     print(product)
@@ -318,7 +352,7 @@ def products_in_store_id():
 
 def products_in_store_name():
     name = input("Enter Store Name: ").title()
-    if name:
+    if name and not name.isdigit():
         departments = Department.find_by_store_name(name)
         if departments:
             store = Store.find_by_name(name)
@@ -331,7 +365,7 @@ def products_in_store_name():
         else:
             print(f"\t>>>> Store with name {name} does not exist <<<<")
     else:
-        print(f"\t>>>> Input cannot be blank <<<<")
+        print(f"\t>>>> Input must be a string <<<<")
 
 def show_almost_out():
     products = Product.get_all()
@@ -339,13 +373,15 @@ def show_almost_out():
         if 0 < product.quantity <= 10:
             department = Department.find_by_id(product.department_id)
             store = Store.find_by_id(department.store_id)
+            print("\n-------------------------Stock Updates-----------------------")
             print(f"{store.name}")
-            print(f"***{product.name} is almost out of stock. Only {product.quantity} remaining!***\n")
+            print(f"***{product.name} is almost out of stock. Only {product.quantity} remaining!***")
+            print("------------------------------------------------------------\n")
         elif product.quantity == 0:
             department = Department.find_by_id(product.department_id)
             store = Store.find_by_id(department.store_id)
             print(f"{store.name}")
-            print(f"***{product.name} is out of stock!!***\n")
+            print(f"***{product.name} is out of stock!!***")
             
             
 def update_quantity():
